move auth and password confs to Auth directory and add ghcoptions in cabal file
