+++ dit-calendar.cabal
@@ -73,6 +73,7 @@ executable dit-calendar
   main-is:             Main.hs
   hs-source-dirs:      src
   default-language:    Haskell2010
+  ghc-options:         -threaded -rtsopts
   build-depends:       base >=4.9 && <4.10,
                        dit-calendar
