+++ library/Auth/ssl/localhost.key
@@ -1,0 +1,15 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIICXQIBAAKBgQCeDjxixdTzrOAEP6BczXhMwPBnjiO/JXkXEirjpWBRlQp/O+3p
+3u/YgvLiaFYxRqZzhGYMefUbnRvhlmzr+f310lLIJwWbZJKu0Q5Stb7OkKf1rKQO
+tTnqShnwO1sDQBZKyPkKM2+VpXXHDx7x08zBn8puGgL2bsyv+KgXiktsHwIDAQAB
+AoGAPhsp70X9X6LWg9TclprGYqKN1KMtuLZo1n/1Cm65YlVGQqix2+wWIuWRuxV0
+iPhR/6JO1yg1ABzED+ij9DXPbGKJaTsjNtHMq9glrlK5nZuSKvYCY6pHC7CWBosu
+c5pqWfqYwHEhU9DpP8JebJuenlkF5oClfpU1JMG10nAeKUECQQDPQja2zvxsSwjG
+3Z5r/9UBt1FZFvzYeQkHEoq8oC28RvmLQOciVsoUTHaUx0fEVjIxAEbjLpcEbNjg
+U5K5V/9bAkEAwznOO1zauZPHtjGYIZXGsEOu+zjVS5aXYliGAomJ9HEIgOSkcJeX
+xa3ND99trbD+YKvHPRKSNPljjXLvzTUFjQJBAMDt9BysBRGRr3PfLO9/0HqueM0j
+tu6TjLHGstrviGNw9i6Ilh/DBo1quB+6ebXsPwCaL2KfJfj6kdAQO87IeW0CQAFw
+VJ/FYGtLtoYWOXvN4cf8x7ZcRd4nvvfflSVBpD2j858XkEqwI1M2o7NBP0qvakCG
+I5LEp0UyrVQmi7PkdyUCQQC6lXt4L0hqd7ftnFIPPzdW5MCuVCNcketDmQia5Ve+
+CSDYM31ZFb6uzYQLBHg30kDA7puBxvUzUDHI1s8Tj8pU
+-----END RSA PRIVATE KEY-----
